requirements.txt
# پیش‌نیازهای پروژه تحلیل گر هوشمند چارت
# برای نصب: pip install -r requirements.txt

# ═══════════════════════════════════════════════════════════════
# 🤖 هوش مصنوعی (یکی را انتخاب کنید)
# ═══════════════════════════════════════════════════════════════

# OpenAI (هزینه‌بر - دقت بالا)
# openai>=1.50.0

# Google Gemini (کاملاً رایگان! - پیشنهادی)
google-generativeai>=0.8.0

# ═══════════════════════════════════════════════════════════════
# 💬 ربات تلگرام
# ═══════════════════════════════════════════════════════════════

aiogram>=3.10.0
aiofiles>=23.2.1

# ═══════════════════════════════════════════════════════════════
# 🌐 سرور Webhook برای TradingView
# ═══════════════════════════════════════════════════════════════

fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# ═══════════════════════════════════════════════════════════════
# 🖼️ پردازش تصویر
# ═══════════════════════════════════════════════════════════════

Pillow>=10.0.0

# ═══════════════════════════════════════════════════════════════
# 🗄️ پایگاه داده (اختیاری)
# ═══════════════════════════════════════════════════════════════

# sqlalchemy>=2.0.0
# asyncpg>=0.29.0

# ═══════════════════════════════════════════════════════════════
# 🛠️ ابزارهای کمکی
# ═══════════════════════════════════════════════════════════════

python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# ═══════════════════════════════════════════════════════════════
# 📊 لاگینگ و مانیتورینگ (اختیاری)
# ═══════════════════════════════════════════════════════════════

# sentry-sdk>=2.0.0

# ═══════════════════════════════════════════════════════════════
# 🧪 تست
# ═══════════════════════════════════════════════════════════════

pytest>=8.0.0

# ═══════════════════════════════════════════════════════════════
# ⏰ زمان‌بندی (اختیاری)
# ═══════════════════════════════════════════════════════════════

apscheduler>=3.10.0
